Move the token verification method to the root package
