+++ api/auth.go
@@ -2,33 +2,15 @@ package api
 
 import (
 	"context"
-	"encoding/base64"
-	"encoding/json"
-	"fmt"
 	"io"
-	"log"
 	"net/http"
 	"os"
-	"strings"
 	"time"
 
 	"github.com/pkg/errors"
 	"github.com/rsp9u/go-oidc"
-	"golang.org/x/oauth2"
 	"minivmm"
 )
-
-const (
-	clientID     = "minivmm"
-	clientSecret = "minivmmminivmm"
-	cookieName   = "minivmm_token"
-)
-
-type jwtPayload struct {
-	Issuer   string `json:"iss"`
-	Subject  string `json:"sub"`
-	ClientID string `json:"client_id"`
-}
 
 // AuthMiddleware is a middleware resolving authentication.
 func AuthMiddleware(next http.Handler) http.Handler {
@@ -50,34 +32,14 @@ func auth(r *http.Request) (bool, *http.Request) {
 		return true, newReq
 	}
 
-	cookie, err := r.Cookie(cookieName)
+	cookie, err := r.Cookie(minivmm.CookieName)
 	if err != nil {
 		return false, r
 	}
 	token := cookie.Value
 
-	// Check client ID and signature of access token
-	payload, err := extractJWTPayload(token)
+	payload, err := minivmm.VerifyToken(token)
 	if err != nil {
-		log.Println("failed to parse access token: ", err)
-		return false, r
-	}
-
-	if payload.ClientID != clientID {
-		log.Println("failed to verify celient ID in access token")
-		return false, r
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-	defer cancel()
-	_, provider, _, err := setupOIDCProvider(ctx)
-	if err != nil {
-		log.Println("failed to setup oidc provider: ", err)
-		return false, r
-	}
-
-	if _, err = provider.RemoteKeySet.VerifySignature(context.Background(), token); err != nil {
-		log.Println("failed to verify signature in access token: ", err)
 		return false, r
 	}
 
@@ -91,7 +53,7 @@ func auth(r *http.Request) (bool, *http.Request) {
 func redirectToAuthURL(w http.ResponseWriter, r *http.Request) {
 	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
 	defer cancel()
-	oauth2Config, _, _, err := setupOIDCProvider(ctx)
+	oauth2Config, _, _, err := minivmm.SetupOIDCProvider(ctx)
 	if err != nil {
 		writeInternalServerError(err, w)
 		return
@@ -117,7 +79,7 @@ func HandleOIDCCallback(w http.ResponseWriter, r *http.Request) {
 	}
 
 	cookie := http.Cookie{
-		Name:   "minivmm_token",
+		Name:   minivmm.CookieName,
 		Value:  accessToken,
 		Path:   "/",
 		Secure: true,
@@ -129,7 +91,7 @@ func HandleOIDCCallback(w http.ResponseWriter, r *http.Request) {
 func generateToken(authCode string) (*oidc.IDToken, string, error) {
 	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
 	defer cancel()
-	oauth2Config, _, verifier, err := setupOIDCProvider(ctx)
+	oauth2Config, _, verifier, err := minivmm.SetupOIDCProvider(ctx)
 	if err != nil {
 		return nil, "", err
 	}
@@ -163,52 +125,4 @@ func generateToken(authCode string) (*oidc.IDToken, string, error) {
 	}
 
 	return idToken, rawAccessToken, nil
-}
-
-func setupOIDCProvider(ctx context.Context) (*oauth2.Config, *oidc.Provider, *oidc.IDTokenVerifier, error) {
-	redirectURL := os.Getenv(minivmm.EnvOrigin) + "/api/v1/login"
-
-	// set up
-	iss := os.Getenv(minivmm.EnvOIDC) + "/"
-	provider, err := oidc.NewProvider(ctx, iss)
-	if err != nil {
-		return nil, nil, nil, errors.New("Failed to set up OIDC provider")
-	}
-	ep := provider.Endpoint()
-	ep.AuthStyle = oauth2.AuthStyleInHeader
-	oauth2Config := &oauth2.Config{
-		ClientID:     clientID,
-		ClientSecret: clientSecret,
-		RedirectURL:  redirectURL,
-		Endpoint:     ep,
-		Scopes:       []string{oidc.ScopeOpenID},
-	}
-	oidcConfig := oidc.Config{
-		ClientID:          clientID,
-		SkipClientIDCheck: false,
-		SkipExpiryCheck:   false,
-		SkipIssuerCheck:   false,
-	}
-	verifier := provider.Verifier(&oidcConfig)
-
-	return oauth2Config, provider, verifier, nil
-}
-
-func extractJWTPayload(jwt string) (*jwtPayload, error) {
-	parts := strings.Split(jwt, ".")
-	if len(parts) < 2 {
-		return nil, fmt.Errorf("malformed jwt, expected 3 parts got %d", len(parts))
-	}
-
-	rawPayload, err := base64.RawURLEncoding.DecodeString(parts[1])
-	if err != nil {
-		return nil, errors.Wrap(err, "malformed jwt payload")
-	}
-
-	var payload jwtPayload
-	if err = json.Unmarshal(rawPayload, &payload); err != nil {
-		return nil, errors.Wrap(err, "failed to unmarshal payload json: "+string(rawPayload))
-	}
-
-	return &payload, nil
 }
